fix: sync multi-row edits to Export
