+++ timeline.html
@@ -74,6 +74,7 @@
 
 <script src="https://cdn.jsdelivr.net/npm/dompurify@3.0.3/dist/purify.min.js"></script>
 <script src="scripts/parseCSV.js"></script>
+<script src="scripts/markdown.js"></script>
 <script src="scripts/timeline.js"></script>
 </body>
-</html> 
+</html>
